update some code
