+++ hudi-client/src/main/java/org/apache/hudi/common/rdd/HoodieRDDBloomIndexV2.java
@@ -3,11 +3,11 @@ package org.apache.hudi.common.rdd;
 import org.apache.hudi.client.WriteStatus;
 import org.apache.hudi.common.HoodieEngineContext;
 import org.apache.hudi.common.HoodieIndexV2;
+import org.apache.hudi.common.HoodieTable;
 import org.apache.hudi.common.HoodieWriteInput;
 import org.apache.hudi.common.HoodieWriteOutput;
 import org.apache.hudi.common.model.HoodieRecord;
 import org.apache.hudi.common.model.HoodieRecordPayload;
-import org.apache.hudi.table.v2.HoodieTableV2;
 import org.apache.spark.api.java.JavaRDD;
 
 public class HoodieRDDBloomIndexV2<T extends HoodieRecordPayload> implements
@@ -22,7 +22,7 @@ public class HoodieRDDBloomIndexV2<T extends HoodieRecordPayload> implements
     @Override
     public HoodieWriteInput<JavaRDD<HoodieRecord<T>>> tagLocation(
         HoodieWriteInput<JavaRDD<HoodieRecord<T>>> inputs, HoodieEngineContext context,
-        HoodieTableV2 table) {
+        HoodieTable table) {
         return null;
     }
