+++ hudi-client/src/main/java/org/apache/hudi/common/rdd/HoodieRDDMergeOnReadTable.java
@@ -18,12 +18,17 @@ import org.apache.hudi.config.HoodieWriteConfig;
 import org.apache.hudi.table.UserDefinedBulkInsertPartitioner;
 import org.apache.hudi.table.action.commit.HoodieWriteMetadata;
 import org.apache.spark.api.java.JavaRDD;
+import org.apache.spark.api.java.JavaSparkContext;
 
 public class HoodieRDDMergeOnReadTable<T extends HoodieRecordPayload> extends HoodieRDDTable<T> {
 
-  public HoodieRDDMergeOnReadTable(HoodieWriteConfig config,
-      HoodieTableMetaClient metaClient) {
-    super(config, metaClient);
+  public HoodieRDDMergeOnReadTable(HoodieWriteConfig config, HoodieTableMetaClient metaClient) {
+    this(new JavaSparkContext(), config, metaClient);
+  }
+
+  public HoodieRDDMergeOnReadTable(JavaSparkContext jsc, HoodieWriteConfig config,
+                                   HoodieTableMetaClient metaClient) {
+    super(jsc, config, metaClient);
   }
 
   @Override
